+++ lib/components/ds_components.dart
@@ -2,5 +2,7 @@ export 'ds_basic_screen_form/ds_basic_screen_form.dart';
 export 'ds_button/ds_button.dart';
 export 'ds_icon_button/ds_icon_button.dart';
 export 'ds_image_view/ds_image_view.dart';
+export 'ds_image_view/image_view_wrapper.dart';
 export 'ds_input/ds_input.dart';
 export 'ds_loading/ds_loading.dart';
+export 'ds_radio/ds_radio.dart';