+++ lib/components/ds_radio/ds_radio.dart
@@ -1,0 +1,84 @@
+import 'package:flutter/material.dart';
+
+import '../../base/ds_base.dart';
+import '../../design_system_core/ds_color_usage/ds_color_usage_core.dart';
+import '../../extensions/extensions.dart';
+import '../../theme/ds_theme.dart';
+
+class DSRadio<T> extends StatefulWidget {
+  /// The value represented by this radio button.
+  final T value;
+
+  /// The currently selected value for a group of radio buttons.
+  ///
+  /// This radio button is considered selected if its [value] matches the
+  /// [groupValue].
+  final T? groupValue;
+
+  /// Called when the user selects this radio button.
+  ///
+  /// The radio button passes [value] as a parameter to this callback. The radio
+  /// button does not actually change state until the parent widget rebuilds the
+  /// radio button with the new [groupValue].
+  ///
+  /// If null, the radio button will be displayed as disabled.
+  ///
+  /// The provided callback will not be invoked if this radio button is already
+  /// selected.
+  ///
+  /// The callback provided to [onChanged] should update the state of the parent
+  /// [StatefulWidget] using the [State.setState] method, so that the parent
+  /// gets rebuilt; for example:
+  ///
+  /// ```dart
+  /// Radio<SingingCharacter>(
+  ///   value: SingingCharacter.lafayette,
+  ///   groupValue: _character,
+  ///   onChanged: (SingingCharacter? newValue) {
+  ///     setState(() {
+  ///       _character = newValue;
+  ///     });
+  ///   },
+  /// )
+  /// ```
+  final ValueChanged<T?>? onChanged;
+
+  const DSRadio({
+    super.key,
+    required this.value,
+    required this.groupValue,
+    required this.onChanged,
+  });
+
+  @override
+  State<DSRadio> createState() => _DSRadioState();
+}
+
+class _DSRadioState extends DSStateBase<DSRadio> {
+  late DSRadioTheme componentTheme =
+      theme.extension<DSRadioThemeExtension>()!.dSRadioTheme;
+
+  bool get isSelected => widget.value == widget.groupValue;
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: widget.onChanged != null
+          ? () => widget.onChanged!(widget.value)
+          : null,
+      child: AnimatedContainer(
+        duration: const Duration(milliseconds: 200),
+        decoration: BoxDecoration(
+          color: colors.white,
+          shape: BoxShape.circle,
+          border: Border.all(
+            color: isSelected
+                ? colors.brand.primary
+                : DSColorUsages.border.primary,
+            width: isSelected ? 8 : 1,
+          ),
+        ),
+      ),
+    );
+  }
+}