+++ lib/theme/base/app_theme/ds_app_theme.dart
@@ -77,6 +77,7 @@ class DSAppTheme {
       DSIconButtonThemeExtension(),
       DSInputThemeExtension(),
       DSBasicScreenFormThemeExtension(_textTheme),
+      DSRadioThemeExtension(),
     ],
   );